requirements.txt
# BigQuery AI Hackathon Dependencies
# Core BigQuery and AI libraries
google-cloud-bigquery>=3.36.0
bigframes>=2.18.0

# Data processing and utilities
pandas>=2.3.2
numpy>=2.3.2
PyYAML>=6.0.1

# Visualization
matplotlib>=3.10.6
seaborn>=0.13.2
plotly>=5.24.1

# Jupyter notebook support
jupyter>=1.1.1
notebook>=7.4.5
pexpect>=4.8.0

# Additional utilities
requests>=2.32.5

# Hugging Face datasets for legal data
datasets>=3.2.0
huggingface-hub>=0.28.1
